pyproject.toml
[project]
name = "captcha-recognize-bundle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.125.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.30.0",
]

[[tool.uv.index]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
